Add files via upload
This program allows users to track and report instances of losing control of their body and muscles, as well as experiences of tripping out. Users can log their experiences with a description, and the program records the date and time of the log entry. Users can also view the log history to see their previous entries.

The logs are stored in a list bodyControlLogs where each entry is a string containing the type of experience (either "Body Control" or "Tripping Out"), the date and time of the log, and the description provided by the user.